add abstract BaseMultiValuesSourceFieldConfig
